Implement BLAST database coding for nucleotide and protein sequences.
For nucleotide sequences this supports several different encodings:
* IUPACNA is an ASCII sequence encoding seen in FASTA files.
* Ncbi8nBase is a 4-bit encoding where each unambiguous base has an
  independent bit. This is used as an intermediate coding and may be
  requested by `BlastSeqSrc` as `eBlastEncodingNcbi4na`.
* Ncbi2nBase is a 2-bit encoding that only allows the 4 unambiguous
  bases. It is used as part of the database encoding and may be
  requested by `BlastSeqSrc` as `eBlastEncodingProtein` (intuitive!).
* Blastna8Base is a 4-bit encoding where unambiguous bases are encoded
  as values 0-3 like Ncbi8naBase but ambiguous bases are encoded using
  different values. This may be requested by `BlastSeqSrc` as
  `eBlastEncodingNucleotide`.

The `NucleotideSequence` struct represents the encoding that appears in
the BLAST database which is split into a packed 2-bit base coding and
a run length coding of ambiguous bases. The former is used during
preliminary search phases, the latter is used for scoring/traceback.

Protein sequences support IUPAC ASCII input and NcbiStdaaBase which fits
all the alphabet values in 5 bits.

Both sequence types support conversion to and from IUPAC strings as well
as accepting partially converted input bases, and both support yielding
byte slices to write to storage.